Changes Home.js to HomeIcon.js, Navigate to MapIcon.js, Settings.js to SettingsIcon.js
